Move rememberObservedNavController out of internal package
